requirements.txt
# Core
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0


# Environment
python-dotenv>=1.0.0

# Build dependencies
setuptools>=70.0.0
wheel>=0.43.0

# Work Libraries